+++ apps/api/src/libs/schema/validators.ts
@@ -3,8 +3,8 @@ import { z } from 'zod';
 import { rpcProviders } from '#libs/near';
 
 const list = z.object({
-  page: z.number().positive().max(200).optional().default(1),
-  per_page: z.number().positive().max(250).optional().default(25),
+  page: z.number().int().positive().max(200).optional().default(1),
+  per_page: z.number().int().positive().max(250).optional().default(25),
   rpc: z.enum(rpcProviders).optional(),
 });
